requirements.txt
ascii-magic=2.3.0
python-telegram-bot==13.11
requests==2.31.0